jira_bus/requirements.txt
restapi-client==1.0.581
auth-client==0.0.79
tornado==6.2
config_client==1.0.100
optscale_exceptions==0.0.21
sqlalchemy==1.4.23
alembic==1.7.1
mysql-connector==2.2.9
apispec[yaml]==5.1.0
PyJWT[crypto]==1.7.1
atlassian_jwt==1.9.0
requests==2.25.0
pydevd-pycharm==201.5616.27
